rename to upstream_wait
