+++ api/envoy/config/filter/http/upstream_wait/v2/upstream_wait_filter.proto
@@ -1,0 +1,10 @@
+syntax = "proto3";
+
+package envoy.config.filter.http.upstream_wait.v2;
+
+option java_package = "io.envoyproxy.envoy.config.filter.http.upstream_wait.v2";
+option java_outer_classname = "UpstreamWaitFilterProto";
+option java_multiple_files = true;
+option go_package = "upstream_wait";
+
+message UpstreamWaitFilterConfig {}
