--- api/envoy/config/filter/http/wait/v2/wait_filter.proto
@@ -1,10 +1,0 @@
-syntax = "proto3";
-
-package envoy.config.filter.http.wait.v2;
-
-option java_package = "io.envoyproxy.envoy.config.filter.http.wait.v2";
-option java_outer_classname = "WaitFilterProto";
-option java_multiple_files = true;
-option go_package = "wait";
-
-message WaitFilterConfig {}
