+++ source/extensions/filters/http/solo_well_known_names.h
@@ -27,7 +27,7 @@ public:
   // TODO(talnordan): Consider "io.solo.filters.http.transformation".
   const std::string Transformation = "io.solo.transformation";
   // Wait filter
-  const std::string Wait = "io.solo.wait";
+  const std::string Wait = "io.solo.upstream_wait";
 };
 
 typedef ConstSingleton<SoloHttpFilterNameValues> SoloHttpFilterNames;
